# Conflicts:
#	src/backend/ci/ext/tencent/dispatch-kubernetes/api-dispatch-kubernetes-tencent/src/main/kotlin/com/tencent/devops/dispatch/kubernetes/pojo/mq/WorkspaceOperateEvent.kt
#	src/backend/ci/ext/tencent/dispatch-kubernetes/biz-dispatch-kubernetes-devcloud-tencent/src/main/kotlin/com/tencent/devops/dispatch/devcloud/service/DevCloudRemoteDevService.kt

+++ src/backend/ci/ext/tencent/process/api-process-tencent/src/main/kotlin/com/tencent/devops/process/api/service/ServiceTXPipelineResource.kt
@@ -73,7 +73,7 @@ interface ServiceTXPipelineResource {
     fun getPipelineInfobyId(
         @PathParam("id")
         @ApiParam(value = "流水线自增id", required = true)
-        id: Int
+        id: Long
     ): Result<SimplePipeline>
 
     @ApiOperation("用模板实例化一条AM流水线并且触发运行")
